+++ tests/testthat/test-odds_ratio.R
@@ -122,7 +122,6 @@ testthat::test_that("estimate_odds_ratio estimates right OR and CI (unstratified
 })
 
 testthat::test_that("estimate_odds_ratio estimates right OR and CI (stratified analysis)", {
-  skip_if_too_deep(5)
   set.seed(12)
   data <- data.frame(
     rsp = sample(c(TRUE, FALSE), 100, TRUE),